+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unpublished
+## 0.4.0
 
 ### Changed